% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blame.R
\name{exo_blame}
\alias{exo_blame}
\title{"I want to find who added a specific line!"}
\usage{
exo_blame(parent_path)
}
\arguments{
\item{parent_path}{Path where to create the exercise repo}
}
\value{
The path to the new project
}
\description{
I want to find who added the line \code{x <- x + 1} to the script and when.
The tool for that is \verb{git blame path/to/file}.
}
\section{Workflow}{

Running the function will create the exercise as a new folder in \code{parent_path}.
If called from RStudio or Positron, the function will open a new R session in
that IDE.
If not, the user will need to navigate to the path returned by the function,
and launch an R session from there.
The new R session will display messages about what challenge to solve,
and running the \code{tip()} function from that new R session will display
additional guiance.
}

\examples{
\dontshow{if (interactive()) withAutoprint(\{ # examplesIf}
parent_path <- withr::local_tempdir()
path <- exo_blame(parent_path = parent_path)
\dontshow{\}) # examplesIf}
}
\section{Related Git documentation}{\href{https://git-scm.com/docs/git-blame}{\code{git blame}}.}

