% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-editor.R
\name{exo_check_editor}
\alias{exo_check_editor}
\title{"Hey, I don't want to get into Vim by mistake!"}
\usage{
exo_check_editor(parent_path)
}
\arguments{
\item{parent_path}{Path where to create the exercise repo}
}
\value{
The path to the new project
}
\description{
An exercice to check your Git's core.editor is set correctly.
\url{https://docs.github.com/es/get-started/git-basics/associating-text-editors-with-git}
}
\section{Workflow}{

Running the function will create the exercise as a new folder in \code{parent_path}.
If called from RStudio or Positron, the function will open a new R session in
that IDE.
If not, the user will need to navigate to the path returned by the function,
and launch an R session from there.
The new R session will display messages about what challenge to solve,
and running the \code{tip()} function from that new R session will display
additional guiance.
}

\examples{
\dontshow{if (interactive()) withAutoprint(\{ # examplesIf}
parent_path <- withr::local_tempdir()
path <- exo_check_editor(parent_path = parent_path)
\dontshow{\}) # examplesIf}
}
\section{Related Git documentation}{\href{https://git-scm.com/docs/git-config}{\code{git config}}, \href{https://git-scm.com/docs/git-commit#Documentation/git-commit.txt---allow-empty}{\code{git commit --allow-empty}}.}

