% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exo-clean-dir.R
\name{exo_clean_dir}
\alias{exo_clean_dir}
\title{"Hey I'd like to remove these untracked files I created to test stuff!"}
\usage{
exo_clean_dir(parent_path)
}
\arguments{
\item{parent_path}{Path where to create the exercise repo}
}
\value{
The path to the new project
}
\description{
If debugging for instance created now useless untracked files and directories,
there's no need to remove them "manually".
The tool for that is \verb{git clean}:
\itemize{
\item \verb{git clean -n} for a dry run;
\item \verb{git clean -f} to run it;
Add \code{-d} to also remove directories.
See \url{https://git-scm.com/docs/git-clean}.
}
}
\section{Workflow}{

Running the function will create the exercise as a new folder in \code{parent_path}.
If called from RStudio or Positron, the function will open a new R session in
that IDE.
If not, the user will need to navigate to the path returned by the function,
and launch an R session from there.
The new R session will display messages about what challenge to solve,
and running the \code{tip()} function from that new R session will display
additional guiance.
}

\examples{
\dontshow{if (interactive()) withAutoprint(\{ # examplesIf}
parent_path <- withr::local_tempdir()
path <- exo_clean_dir(parent_path = parent_path)
\dontshow{\}) # examplesIf}
}
\section{Related Git documentation}{\href{https://git-scm.com/docs/git-clean}{\code{git clean}}.}

