% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exo-conflict.R
\name{exo_conflict}
\alias{exo_conflict}
\title{"Hey I'd like to see what merge conflicts look like!"}
\usage{
exo_conflict(parent_path)
}
\arguments{
\item{parent_path}{Path where to create the exercise repo}
}
\value{
The path to the new project
}
\description{
I made some work in a feature branch and want to merge it.
Meanwhile, the main branch advanced.
Unfortunately someone touched the same file as I did.
Now I need to fix a merge conflict!

See also \url{https://happygitwithr.com/git-branches.html#dealing-with-conflicts}.
}
\section{Workflow}{

Running the function will create the exercise as a new folder in \code{parent_path}.
If called from RStudio or Positron, the function will open a new R session in
that IDE.
If not, the user will need to navigate to the path returned by the function,
and launch an R session from there.
The new R session will display messages about what challenge to solve,
and running the \code{tip()} function from that new R session will display
additional guiance.
}

\examples{
\dontshow{if (interactive()) withAutoprint(\{ # examplesIf}
parent_path <- withr::local_tempdir()
path <- exo_conflict(parent_path = parent_path)
\dontshow{\}) # examplesIf}
}
\section{Related Git documentation}{\href{https://git-scm.com/docs/git-merge}{\code{git merge}}.}

