% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exo-log-deleted-file.R
\name{exo_log_deleted_file}
\alias{exo_log_deleted_file}
\title{"I want to find which commit deleted a file!"}
\usage{
exo_log_deleted_file(parent_path)
}
\arguments{
\item{parent_path}{Path where to create the exercise repo}
}
\value{
The path to the new project
}
\description{
I want to find which commit deleted script.R.
The tool for that is \verb{git log --oneline -- path}.
}
\section{Workflow}{

Running the function will create the exercise as a new folder in \code{parent_path}.
If called from RStudio or Positron, the function will open a new R session in
that IDE.
If not, the user will need to navigate to the path returned by the function,
and launch an R session from there.
The new R session will display messages about what challenge to solve,
and running the \code{tip()} function from that new R session will display
additional guiance.
}

\examples{
\dontshow{if (interactive()) withAutoprint(\{ # examplesIf}
parent_path <- withr::local_tempdir()
path <- exo_log_deleted_file(parent_path = parent_path)
\dontshow{\}) # examplesIf}
}
\section{Related Git documentation}{\href{https://git-scm.com/docs/git-log#Documentation/git-log.txt---path}{\code{git log -- <path>}}.}

