% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/one-small-change.R
\name{exo_one_small_change}
\alias{exo_one_small_change}
\title{"Oh shit, I committed and immediately realized I need to make one small change!"}
\usage{
exo_one_small_change(parent_path)
}
\arguments{
\item{parent_path}{Path where to create the exercise repo}
}
\value{
The path
}
\description{
To go with \url{https://ohshitgit.com/#change-last-commit}
}
\section{Workflow}{

Running the function will create the exercise as a new folder in \code{parent_path}.
If called from RStudio or Positron, the function will open a new R session in
that IDE.
If not, the user will need to navigate to the path returned by the function,
and launch an R session from there.
The new R session will display messages about what challenge to solve,
and running the \code{tip()} function from that new R session will display
additional guiance.
}

\examples{
\dontshow{if (interactive()) withAutoprint(\{ # examplesIf}
parent_path <- withr::local_tempdir()
path <- exo_one_small_change(parent_path = parent_path)
# Now add "thing 3" to the "bla" file
# And amend the latest commit
\dontshow{\}) # examplesIf}
}
\section{Related Git documentation}{\href{https://git-scm.com/docs/git-commit#Documentation/git-commit.txt---no-edit}{\code{git commit --amend --no-edit}}.}

