% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exo-split-changes.R
\name{exo_split_changes}
\alias{exo_split_changes}
\title{"Hey I'd like to split these changes to the same file into several commits!"}
\usage{
exo_split_changes(parent_path)
}
\arguments{
\item{parent_path}{Path where to create the exercise repo}
}
\value{
The path to the new project
}
\description{
I made many edits to a file, in different places.
This is too much for a commit, since small commits are best practice.
I need to add the changes to Git bit by bit.
The tool for that is \verb{git add --patch}, also available as \verb{git add -p}.
If all your changes are presented to you as one chunk by \verb{git add --patch},
choose the "s" option for splitting.
See \url{https://git-scm.com/book/en/v2/Git-Tools-Interactive-Staging#_staging_patches}.

Note that patch is also an option for \verb{git commit}, if you prefer so.
}
\section{Workflow}{

Running the function will create the exercise as a new folder in \code{parent_path}.
If called from RStudio or Positron, the function will open a new R session in
that IDE.
If not, the user will need to navigate to the path returned by the function,
and launch an R session from there.
The new R session will display messages about what challenge to solve,
and running the \code{tip()} function from that new R session will display
additional guiance.
}

\examples{
\dontshow{if (interactive()) withAutoprint(\{ # examplesIf}
parent_path <- withr::local_tempdir()
path <- exo_split_changes(parent_path = parent_path)
\dontshow{\}) # examplesIf}
}
\section{Related Git documentation}{\href{https://git-scm.com/docs/git-add#Documentation/git-add.txt---patch}{\code{git add --patch}}.}

