% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/one-small-change.R
\name{one_small_change}
\alias{one_small_change}
\title{"Oh shit, I committed and immediately realized I need to make one small change!"}
\usage{
one_small_change(path)
}
\arguments{
\item{path}{Path where to create the exercise repo}
}
\value{
The path
}
\description{
"Oh shit, I committed and immediately realized I need to make one small change!"
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
path <- withr::local_tempdir()
one_small_change(path = path)
# Now add "thing 3" to "bla"
# And amend the latest commit
\dontshow{\}) # examplesIf}
}
