% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saperlipopette-package.R
\docType{package}
\name{saperlipopette-package}
\alias{saperlipopette}
\alias{saperlipopette-package}
\title{saperlipopette: Create Example Git Messes}
\description{
Holds functions creating Git messes, that users would then solve, to follow \url{https://ohshitgit.com/}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://docs.ropensci.org/saperlipopette/}
  \item \url{https://github.com/ropensci-training/saperlipopette}
  \item Report bugs at \url{https://github.com/ropensci-training/saperlipopette/issues}
}

}
\author{
\strong{Maintainer}: Maëlle Salmon \email{msmaellesalmon@gmail.com} (\href{https://orcid.org/0000-0002-2815-0399}{ORCID})

Authors:
\itemize{
  \item Maëlle Salmon \email{msmaellesalmon@gmail.com} (\href{https://orcid.org/0000-0002-2815-0399}{ORCID})
  \item Yanina Bellini Saibene (\href{https://orcid.org/0000-0002-4522-7466}{ORCID})
}

Other contributors:
\itemize{
  \item Jim Gardner [contributor]
  \item Steffi LaZerte (\href{https://orcid.org/0000-0002-7690-8360}{ORCID}) (Steffi reviewed the 'saperlipopette, a risk-free playground for learning more Git' article.) [reviewer]
}

}
\keyword{internal}
