% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create-all.R
\name{create_all_exercises}
\alias{create_all_exercises}
\title{Create all exercises folder at once}
\usage{
create_all_exercises(parent_path)
}
\arguments{
\item{parent_path}{Path where to create the exercise repo}
}
\value{
The parent path
}
\description{
But do not open them!
}
