% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create-all.R
\name{create_all_exercises}
\alias{create_all_exercises}
\title{Create all exercises folder at once}
\usage{
create_all_exercises(parent_path)
}
\arguments{
\item{parent_path}{Path where to create the exercise repo}
}
\value{
The parent path
}
\description{
But do not open them!
}
\section{Workflow}{

Running the function will create the exercise as a new folder in \code{parent_path}.
If called from RStudio or Positron, the function will open a new R session in
that IDE.
If not, the user will need to navigate to the path returned by the function,
and launch an R session from there.
The new R session will display messages about what challenge to solve,
and running the \code{tip()} function from that new R session will display
additional guiance.
}

\examples{
\dontshow{if (interactive()) withAutoprint(\{ # examplesIf}
parent_path <- withr::local_tempdir()
path <- create_all_exercises(parent_path = parent_path)
\dontshow{\}) # examplesIf}
}
