% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/debug.R
\name{exo_bisect}
\alias{exo_bisect}
\title{"Hey I'd like to find which commit introduced a bug!"}
\usage{
exo_bisect(parent_path)
}
\arguments{
\item{parent_path}{Path where to create the exercise repo}
}
\value{
The path to the new project
}
\description{
I notice a bug in my codebase.
I can see the bug was not there a bunch of commits ago.
Beside doing regular debugging, I can find out
which commit introduced the bug by using \verb{git bisect}.
See \url{https://git-scm.com/docs/git-bisect} and
\url{https://www.jimhester.com/post/2019-04-24-git-bisect/}.
}
\section{Git commands}{

\verb{git bisect}.
}

\examples{
\dontshow{if (interactive()) withAutoprint(\{ # examplesIf}
parent_path <- withr::local_tempdir()
path <- exo_bisect(parent_path = parent_path)
\dontshow{\}) # examplesIf}
}
