% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-editor.R
\name{exo_check_editor}
\alias{exo_check_editor}
\title{"Hey, I don't want to get into Vim by mistake!"}
\usage{
exo_check_editor(parent_path)
}
\arguments{
\item{parent_path}{Path where to create the exercise repo}
}
\value{
The path to the new project
}
\description{
An exercice to check your Git's core.editor is set correctly.
}
\section{Git commands}{

\verb{git config}, \verb{git commit --allow-empty}
}

\examples{
\dontshow{if (interactive()) withAutoprint(\{ # examplesIf}
parent_path <- withr::local_tempdir()
path <- exo_check_editor(parent_path = parent_path)
\dontshow{\}) # examplesIf}
}
