% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/committed-to-main.R
\name{exo_committed_to_main}
\alias{exo_committed_to_main}
\title{"Oh shit, I accidentally committed something to main that should have been on a brand new branch!"}
\usage{
exo_committed_to_main(parent_path)
}
\arguments{
\item{parent_path}{Path where to create the exercise repo}
}
\value{
The path to the new project
}
\description{
To go with \url{https://ohshitgit.com/#accidental-commit-master}
}
\examples{
\dontshow{if (interactive()) withAutoprint(\{ # examplesIf}
parent_path <- withr::local_tempdir()
path <- exo_committed_to_main(parent_path = parent_path)
\dontshow{\}) # examplesIf}
}
\section{Related Git documentation}{\href{https://git-scm.com/docs/git-reset#Documentation/git-reset.txt---hard}{\code{git reset --hard}}, \href{https://git-scm.com/docs/git-branch}{\code{git branch}}, \href{https://git-scm.com/docs/git-switch}{\code{git switch}}, \href{https://git-scm.com/docs/git-checkout}{\code{git checkout}}.}

