% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conflict.R
\name{exo_conflict}
\alias{exo_conflict}
\title{"Hey I'd like to see what merge conflicts look like!"}
\usage{
exo_conflict(parent_path)
}
\arguments{
\item{parent_path}{Path where to create the exercise repo}
}
\value{
The path to the new project
}
\description{
I made some work in a feature branch and want to merge it.
Meanwhile, the main branch advanced.
Unfortunately someone touched the same file as I did.
Now I need to fix a merge conflict!

See also \url{https://happygitwithr.com/git-branches.html#dealing-with-conflicts}.
}
\section{Git commands}{

\verb{git merge}.
}

\examples{
\dontshow{if (interactive()) withAutoprint(\{ # examplesIf}
parent_path <- withr::local_tempdir()
path <- exo_conflict(parent_path = parent_path)
\dontshow{\}) # examplesIf}
}
