% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edit-latest-message.R
\name{exo_latest_message}
\alias{exo_latest_message}
\title{"Oh shit, I need to change the message on my last commit!"}
\usage{
exo_latest_message(parent_path)
}
\arguments{
\item{parent_path}{Path where to create the exercise repo}
}
\value{
The path
}
\description{
To go with \url{https://ohshitgit.com/#change-last-commit-message}
}
\section{Git commands}{

\verb{git commit --amend}
}

\examples{
\dontshow{if (interactive()) withAutoprint(\{ # examplesIf}
parent_path <- withr::local_tempdir()
path <- exo_latest_message(parent_path = parent_path)
\dontshow{\}) # examplesIf}
}
