% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exo-latest-message.R
\name{exo_latest_message}
\alias{exo_latest_message}
\title{"Oh shit, I need to change the message on my last commit!"}
\usage{
exo_latest_message(parent_path)
}
\arguments{
\item{parent_path}{Path where to create the exercise repo}
}
\value{
The path
}
\description{
To go with \url{https://ohshitgit.com/#change-last-commit-message}
}
\section{Workflow}{

Running the function will create the exercise as a new folder in \code{parent_path}.
If called from RStudio or Positron, the function will open a new R session in
that IDE.
If not, the user will need to navigate to the path returned by the function,
and launch an R session from there.
The new R session will display messages about what challenge to solve,
and running the \code{tip()} function from that new R session will display
additional guiance.
}

\examples{
\dontshow{if (interactive()) withAutoprint(\{ # examplesIf}
parent_path <- withr::local_tempdir()
path <- exo_latest_message(parent_path = parent_path)
\dontshow{\}) # examplesIf}
}
\section{Related Git documentation}{\href{https://git-scm.com/docs/git-commit#Documentation/git-commit.txt---amend}{\code{git commit --amend}}.}

