% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log-deleted-line.R
\name{exo_log_deleted_line}
\alias{exo_log_deleted_line}
\title{"I want to find which commit deleted a line!"}
\usage{
exo_log_deleted_line(parent_path)
}
\arguments{
\item{parent_path}{Path where to create the exercise repo}
}
\value{
The path to the new project
}
\description{
I want to find which commit deleted the line "iris" from the text file.
The tool for that is \verb{git log -S<string> path/to/file}
or \verb{git log -G<regex> path/to/file}.
}
\examples{
\dontshow{if (interactive()) withAutoprint(\{ # examplesIf}
parent_path <- withr::local_tempdir()
path <- exo_log_deleted_line(parent_path = parent_path)
\dontshow{\}) # examplesIf}
}
\section{Related Git documentation}{\href{https://git-scm.com/docs/git-log#Documentation/git-log.txt--Sstring}{\code{git log -S<string>}}, \href{https://git-scm.com/docs/git-log#Documentation/git-log.txt--Gregex}{\code{git log -G<regex>}}.}

