% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rebase-i.R
\name{exo_rebase_i}
\alias{exo_rebase_i}
\title{"Hey I'd like to make my commits in a branch look informative and smart!"}
\usage{
exo_rebase_i(parent_path)
}
\arguments{
\item{parent_path}{Path where to create the exercise repo}
}
\value{
The path to the new project
}
\description{
I am working in a feature branch that's all my own.
I made many small commits as I was figuring things out.
Now I want the commits to tell a story for the PR reviewers,
and not a story of how many stupid mistakes I made!
The tool for that is \verb{git base --interactive} also available as \verb{git rebase -i}.
Useful links:
\itemize{
\item \url{https://jvns.ca/blog/2023/11/06/rebasing-what-can-go-wrong-/}
\item \url{https://wizardzines.com/comics/rules-for-rebasing/}
\item \url{https://github.com/MikeMcQuaid/GitInPractice/blob/main/06-RewritingHistoryAndDisasterRecovery.adoc#rebase-commits-interactively-git-rebase-interactive}
}
}
\section{Git commands}{

\verb{git rebase -i}
}

\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
parent_path <- withr::local_tempdir()
path <- exo_rebase_i(parent_path = parent_path)
\dontshow{\}) # examplesIf}
}
