% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reset.R
\name{exo_reset}
\alias{exo_reset}
\title{"Hey I'd like to restart from scratch"}
\usage{
exo_reset(parent_path)
}
\arguments{
\item{parent_path}{Path where to create the exercise repo}
}
\value{
The path to the new project
}
\description{
I am working in a feature branch that's all my own.
I made many small commits as I was figuring things out.
Now I want the commits to tell a story for the PR reviewers,
and not a story of how many stupid mistakes I made!
Instead of \verb{git base --interactive} also available as \verb{git rebase -i},
I can also use \verb{git reset --mixed} and then build the commits.
Useful links:
\itemize{
\item \url{https://github.blog/developer-skills/github/write-better-commits-build-better-projects/}
\item \url{https://masalmon.eu/2024/06/11/rewrite-git-history/}
}
}
\section{Workflow}{

Running the function will create the exercise as a new folder in \code{parent_path}.
If called from RStudio or Positron, the function will open a new R session in
that IDE.
If not, the user will need to navigate to the path returned by the function,
and launch an R session from there.
The new R session will display messages about what challenge to solve,
and running the \code{tip()} function from that new R session will display
additional guiance.
}

\examples{
\dontshow{if (interactive()) withAutoprint(\{ # examplesIf}
parent_path <- withr::local_tempdir()
path <- exo_reset(parent_path = parent_path)
\dontshow{\}) # examplesIf}
}
\section{Related Git documentation}{\href{https://git-scm.com/docs/git-reset#Documentation/git-reset.txt---mixed}{\code{git reset --mixed}}.}

