% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exo-revert-file.R
\name{exo_revert_file}
\alias{exo_revert_file}
\title{"Oh shit, I need to undo my changes to a file!"}
\usage{
exo_revert_file(parent_path)
}
\arguments{
\item{parent_path}{Path where to create the exercise repo}
}
\value{
The path to the new project
}
\description{
To go with \url{https://ohshitgit.com/#undo-a-file}
}
\section{Workflow}{

Running the function will create the exercise as a new folder in \code{parent_path}.
If called from RStudio or Positron, the function will open a new R session in
that IDE.
If not, the user will need to navigate to the path returned by the function,
and launch an R session from there.
The new R session will display messages about what challenge to solve,
and running the \code{tip()} function from that new R session will display
additional guiance.
}

\examples{
\dontshow{if (interactive()) withAutoprint(\{ # examplesIf}
parent_path <- withr::local_tempdir()
path <- exo_revert_file(parent_path = parent_path)
\dontshow{\}) # examplesIf}
}
\section{Related Git documentation}{\href{https://git-scm.com/docs/git-log}{\code{git log}}, \href{https://git-scm.com/docs/git-restore}{\code{git restore}}, \href{https://git-scm.com/docs/git-checkout}{\code{git checkout}}, \href{https://git-scm.com/docs/git-commit}{\code{git commit}}.}

