% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/revert-file.R
\name{exo_revert_file}
\alias{exo_revert_file}
\title{"Oh shit, I need to undo my changes to a file!"}
\usage{
exo_revert_file(parent_path)
}
\arguments{
\item{parent_path}{Path where to create the exercise repo}
}
\value{
The path to the new project
}
\description{
To go with \url{https://ohshitgit.com/#undo-a-file}
}
\examples{
\dontshow{if (interactive()) withAutoprint(\{ # examplesIf}
parent_path <- withr::local_tempdir()
path <- exo_revert_file(parent_path = parent_path)
\dontshow{\}) # examplesIf}
}
\section{Related Git documentation}{\href{https://git-scm.com/docs/git-log}{\code{git log}}, \href{https://git-scm.com/docs/git-restore}{\code{git restore}}, \href{https://git-scm.com/docs/git-checkout}{\code{git checkout}}, \href{https://git-scm.com/docs/git-commit}{\code{git commit}}.}

