% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/revparse.R
\name{exo_revparse}
\alias{exo_revparse}
\title{"I want to understand ancestry references!"}
\usage{
exo_revparse(parent_path)
}
\arguments{
\item{parent_path}{Path where to create the exercise repo}
}
\value{
The path to the new project
}
\description{
I want to find the commit ID and message for HEAD~5 and HEAD^^.
The tool for that is \verb{git rev-parse} combined with \verb{git log} or \verb{git show}.
}
\examples{
\dontshow{if (interactive()) withAutoprint(\{ # examplesIf}
parent_path <- withr::local_tempdir()
path <- exo_revparse(parent_path = parent_path)
\dontshow{\}) # examplesIf}
}
\section{Related Git documentation}{\href{https://git-scm.com/docs/git-rev-parse}{\code{git rev-parse}}, \href{https://git-scm.com/docs/git-log}{\code{git log}}, \href{https://git-scm.com/docs/git-show}{\code{git show}}.}

