% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time-machine.R
\name{exo_time_machine}
\alias{exo_time_machine}
\title{"Oh shit, I did something terribly wrong, please tell me git has a magic time machine!?!"}
\usage{
exo_time_machine(parent_path)
}
\value{
The path to the new project
}
\description{
To go with \url{https://ohshitgit.com/#magic-time-machine}
}
\section{Git commands}{

\verb{git reset}, \verb{git reflog}
}

\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
parent_path <- withr::local_tempdir()
path <- exo_time_machine(parent_path = parent_path)
\dontshow{\}) # examplesIf}
}
