% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exo-time-machine.R
\name{exo_time_machine}
\alias{exo_time_machine}
\title{"Oh shit, I did something terribly wrong, please tell me git has a magic time machine!?!"}
\usage{
exo_time_machine(parent_path)
}
\arguments{
\item{parent_path}{Path where to create the exercise repo}
}
\value{
The path to the new project
}
\description{
To go with \url{https://ohshitgit.com/#magic-time-machine}
}
\section{Workflow}{

Running the function will create the exercise as a new folder in \code{parent_path}.
If called from RStudio or Positron, the function will open a new R session in
that IDE.
If not, the user will need to navigate to the path returned by the function,
and launch an R session from there.
The new R session will display messages about what challenge to solve,
and running the \code{tip()} function from that new R session will display
additional guiance.
}

\examples{
\dontshow{if (interactive()) withAutoprint(\{ # examplesIf}
parent_path <- withr::local_tempdir()
path <- exo_time_machine(parent_path = parent_path)
\dontshow{\}) # examplesIf}
}
\section{Related Git documentation}{\href{https://git-scm.com/docs/git-reset}{\code{git reset}}, \href{https://git-scm.com/docs/git-reflog}{\code{git reflog}}.}

