% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/revert-commit.R
\name{exo_undo_commit}
\alias{exo_undo_commit}
\title{"Oh shit, I need to undo a commit from like 5 commits ago!"}
\usage{
exo_undo_commit(parent_path)
}
\arguments{
\item{parent_path}{Path where to create the exercise repo}
}
\value{
The path to the new project
}
\description{
To go with \url{https://ohshitgit.com/#undo-a-commit}
}
\section{Git commands}{

\verb{git log}, \verb{git revert}
}

\examples{
\dontshow{if (interactive()) withAutoprint(\{ # examplesIf}
parent_path <- withr::local_tempdir()
path <- exo_undo_commit(parent_path = parent_path)
\dontshow{\}) # examplesIf}
}
