% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exo-worktree.R
\name{exo_worktree}
\alias{exo_worktree}
\title{"I need to see what the project looked like at a certain version!"}
\usage{
exo_worktree(parent_path)
}
\arguments{
\item{parent_path}{Path where to create the exercise repo}
}
\value{
The path to the new project
}
\description{
I want to create a folder containing the project as it was at tag v2,
to read the text file.
The tool for that is \verb{git worktree}.
\url{https://masalmon.eu/2024/01/23/git-worktree/}
}
\section{Workflow}{

Running the function will create the exercise as a new folder in \code{parent_path}.
If called from RStudio or Positron, the function will open a new R session in
that IDE.
If not, the user will need to navigate to the path returned by the function,
and launch an R session from there.
The new R session will display messages about what challenge to solve,
and running the \code{tip()} function from that new R session will display
additional guiance.
}

\examples{
\dontshow{if (interactive()) withAutoprint(\{ # examplesIf}
parent_path <- withr::local_tempdir()
path <- exo_worktree(parent_path = parent_path)
\dontshow{\}) # examplesIf}
}
\section{Related Git documentation}{\href{https://git-scm.com/docs/git-log}{\code{git log}}, \href{https://git-scm.com/docs/git-worktree}{\code{git worktree}}, \href{https://git-scm.com/docs/git-tag}{\code{git tag}}.}

