% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/one-small-change.R
\name{one_small_change}
\alias{one_small_change}
\title{"Oh shit, I committed and immediately realized I need to make one small change!"}
\usage{
one_small_change(parent_path)
}
\arguments{
\item{parent_path}{Path where to create the exercise repo}
}
\value{
The path
}
\description{
To go with \url{https://ohshitgit.com/#change-last-commit}
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
parent_path <- withr::local_tempdir()
path <- one_small_change(parent_path = parent_path)
fs::dir_tree(path)
gert::git_log(repo = path)
# Now add "thing 3" to "bla"
# And amend the latest commit
\dontshow{\}) # examplesIf}
}
